v0.5b
Reworked how the building block classes handle OnEnable/Start/Init specific code , cutting out the long strings of base.OnEnable and similar call to make things more clear via explicit enable methods (the higher level derived classes still use base.OnEnable() )

Changed the TakeDmg method to accept an int for the type of damage, making it easier to define your own DmgType or equivalent enum then use it in  your own derived TakeDmg methods.

+++ Core/AI/BhvTree/Branches/BNpcAtk.cs
@@ -71,7 +71,7 @@ namespace Hubris
 						if ( b.ActionReady )
 						{
 							a.PlaySound( SndT.ATK );
-							a.TargetEnt.TakeDmg( a, a.Params.DamageStats.CommonType, a.Params.DamageStats.CommonAmount, false );
+							a.TargetEnt.TakeDmg( a, (int)a.Params.DamageStats.CommonType, a.Params.DamageStats.CommonAmount, false );
 						}
 					}
